Treat fields with @AddToRuleKey(stringify = true) as excluded for inputs/deps/serialization
Summary:
If a field uses stringify=true, then its rulekey is not completely
reflecting what that field holds. This change makes inputs/deps derivation and
serialization treat that field as though it is not added to the rulekey (and
the user needs to use CustomFieldBehavior to get the behavior that they
want).

This basically means that inputs and deps derivation are deriving from the
stringified version just like rulekeys (as strings have no inputs/deps). And the only
thing that we know how to deserialize from a string would be a string, and so the
only reasonable way to handle such a field for serialization would be to reject it.

Reviewed By: ttsugriy

fbshipit-source-id: 62625c9