rust: don't use `BuildRule(Params).getBuildDeps()`
Summary:
Avoid using `BuildRule.getBuildDeps()` and `BuildRuleParams.getBuildDeps()` in
Rust descriptions and replace them with `CxxDeps` objects.  Note that despite
the name, `CxxDeps` isn't C/C++ specific -- and will be fixed soon to just take
a flavor instead of a `CxxPlatform` -- and is just a helper to build dep lists
that works with platform-specific deps.

As part of this diff, a new deps method is added to `RustLinkable` to avoid
using `BuildRule.getBuildDeps()`.

Reviewed By: styurin

fbshipit-source-id: 53ef946