add 'extra_shallow_targets' arg to xcode_workspace
Summary: 'extra_targets' does a recursive add of the targets, and it is a simple array, so no options can be added to the targets.  this adds a simple way to provide simple targets to a scheme as a non-recursive addition.  practically speaking, this can be used to add the containing app target to an apple watch target, as the containing app must be built as part of the watch app scheme.  recursive explicit dependencies do not work in this case, so we need a shallow include.

Reviewed By: AnuiShark

fbshipit-source-id: d00fefc