Abstract underlying serialization protocol
Summary:
Right now, a lot of the implementation of different build strategies is
using and creating the generated Thrift types directly. This change
abstracts away the protocol and some logic around computing digests
out to an interface.

Currently, the only implementation of that interface is the ThriftProtocol.
This also consolidates a lot of the use of ThriftProtocol.COMPACT and
specification of the hashing function into the protocol.

Reviewed By: ttsugriy

fbshipit-source-id: 4b2d54f