Rewrite InputsDigestBuilder and rename to FileTreeBuilder
Summary:
InputsDigestBuilder is actually used for both inputs and outputs. Also, it
mixes together logic around constructing the actual merkle tree digests
and such with tracking of the input files in a rather complicated and brittle
way.

This change renames it to FileTreeBuilder and more explicitly separates
the logic into different parts.

Reviewed By: ttsugriy

fbshipit-source-id: 122a254