Acquire command through the CAS
Summary:
In remote execution, the command itself should be passed via the CAS
instead of in the request directly. This allows the command to be reused
across changes to inputs the don't affect the command (which is rather
common).

This changes RemoteExecution to push the command into the CAS
and to acquire it from the CAS.

Reviewed By: ttsugriy

fbshipit-source-id: b2dad07