Add MultiThreadedBlobUploader
Summary:
Currently, the logic for uploading files to the CAS is all contained
within LocalContentAddressedStorage. Also, it's written in such
a way that is not really compatible with a remote CAS (i.e. it
assumes that checking existence of a file is relatively fast).

This change introduces MultiThreadedBlobUploader. This
class accepts requests to upload files and adds them to a
queue for the "missing check". Work threads will pull up to
 missingCheckLimit items off this queue and send a request to
the CAS to find which it does/doesn't contain. Any that are missing
will be added to a queue to be uploaded. Work threads
will then pull those off and upload them. When the upload is finished,
the future for that digest will be fulfilled.

The BlobUploader delegates to a simple interface for the actual missing
check and files upload. This will allow us to easily implement in terms
of various different CAS implementations.

Reviewed By: ttsugriy

fbshipit-source-id: f11994d