Extract InputsMaterializer
Summary:
Like was the case for outputs materialization and inputs uploading, the
 logic for inputs materialization is currently implemented only in
LocalContentAddressedStorage. This extracts it to its own class with a
simple Delegate for the specific implementation.

As inputs materialization is not actually done in the client, this removes
the method from the base CAS interface.

I'm not sure if we'll actually need other implementations of this or not
yet, so I've left it as an inner class of LocalContentAddressedStorage.

Reviewed By: ttsugriy

fbshipit-source-id: ee85274