Use temporary directory as a root of test cell
Summary:
It seems that the current logic with the test libraries is flaky with multiple tests executing concurrently.

This change modifies the logic in integration tests to create a temporary directory for a root of a repository with test libraries and to link all the directories in test library directory to that repository. This will separate test library cells in individual tests.

Reviewed By: bobyangyf

fbshipit-source-id: 8e07887