Fix assert when building target normally after building compilation database
Summary:
`IncrementalActionGraphGenerator::shouldInvalidateParentChain` was relying on `TargetGraph::getOptional` to check if a node with a given target exists in the last target graph; however, `VersionedTargetGraph` overrides `TargetGraph::getInternal`, and can return an unflavored (or worse, differently flavored) version of a node, which breaks the assumption in the assert in `shouldInvalidateParentChain`.

The fix is to add an explicit `getExactOptional` method to `TargetGraph` that does what we want.

Reviewed By: bobyangyf

fbshipit-source-id: 7af0902