Specify glob_includedotfiles when doing a Watchman glob.
Summary:
We need to make sure that the value of `include_dotfiles` passed to a Buck
`glob()` makes its way down to the implementation of `glob()` within Watchman
and is honored accordingly. Previous to this change, it appears that if both
`watchman_use_glob_generator` and `excludes` are `True` when
`format_watchman_query_params()` is called, then the value of
`include_dotfiles` has no effect on the query that is passed to Watchman.

Fortunately, Watchman has an undocumented API (though there is a test for it!)
where specifying `glob_includedotfiles=true` on a `"glob"` query
(https://facebook.github.io/watchman/docs/file-query.html#glob-generator) will
set an internal flag (`WM_PERIOD`) on the `w_query` object:

https://github.com/facebook/watchman/blob/v4.9.0/query/glob.cpp#L159-L165

Therefore, we update `format_watchman_query_params()` so that when
`watchman_use_glob_generator` is `True`, we always make sure to set the
`glob_includedotfiles` property on the Watchman query to match the value of
`include_dotfiles`.

(Incidentally, wez believes that the `watchman_use_glob_generator` boolean
is present to support Watchman <4.7 that did not support `glob` queries, though
we should be able to eliminate that now, which would clean up a bit of code here.
That is work for a separate revision.)

Logic to make use of this flag will be done in a separate Watchman change.
In particular, we need to make sure we make use of this value in the Eden watcher,
which is definitely not happening today:

https://github.com/facebook/watchman/blob/v4.9.0/watcher/eden.cpp

Fortunately, the `w_query` is available where the Thrift calls to Eden's `glob()`
method are made in that file, so it should not be difficult to fix.

Reviewed By: ttsugriy

fbshipit-source-id: f9b4619