Remove references to other BuildRules
Summary:
This passes around the precompiled header sourcepath and whether it is
actually precompiled instead of a reference to the entire pch build rule in
cases where that is appropriate.

Referencing a BuildRule is something that we should strive to avoid, especially
when all we need it for is to find its output file.

Reviewed By: ttsugriy

fbshipit-source-id: 1e0eba5