Make AbstractHeaderVerification use AddsToRuleKey
Summary:
Makes this use AddToRuleKey annotations instead of a custom
appendToRuleKey method(). This slightly changes the logic such that even if the
mode is IGNORE, the whitelist will be added to the rulekey, but I don't expect
that to negatively affect anyone (it would be strange to be ignoring the
verification and building with multiple different whitelists).

Reviewed By: jtorkkola

fbshipit-source-id: 94a537e