Make PreprocessDelegate handle rulekey for the sandbox
Summary:
For sandboxed sources, currently PreprocessDelegate handles all the
logic around the sandboxing, but CxxPreprocessAndCompile handles the logic for
computing the rulekey. It's much better for PreprocessDelegate to properly
compute a rulekey based on its own behavior (and not rely on users to know that
they need to reach into it to change their rulekeys). This moves the rulekey
logic into PreprocessDelegate and removes the SandboxTree field from
CxxPreprocessAndCompile.

Reviewed By: jtorkkola

fbshipit-source-id: a0900d6