Make CxxPreprocessAndCompile derive its own deps
Summary:
One tricky thing about CxxPreprocessAndCompile's dependencies is that
a bunch of them are aggregated into a DependencyAggregation rule. The rules
aggregated in this way are a subset of the dependencies needed by the
PreprocessDelegate (but not all of them). This change moves that dependency
aggregation to be explicitly handled by PreprocessDelegate and makes
CxxPreprocessAndCompile use BuildableSupport to derive its deps.

Reviewed By: jtorkkola

fbshipit-source-id: 18efe4d