Make CxxFlags only capture the part of the platform that it needs
Summary:
CxxFlags only needs the sanitizer, not the entire platform so this
makes it capture that instead. Also, as the sanitizer changes the behavior of
this class, it's added to the rulekey (in practice, this sanitizer instance
will be shared across a ton of rules and so this will just be a cache
lookup when adding it).

Reviewed By: ttsugriy

fbshipit-source-id: a596a11