Add more tests for glob().
Summary:
Note that the way the tests are set up today, this only verifies our
custom implementation of `glob()` in Python (which is built on top of
Python 3's pathlib with a few tweaks).

We need to make sure that `glob()` behaves the same way when backed by
Watchman, as well as however it is implemented when Skylark is used to
process build files.

Reviewed By: ttsugriy

fbshipit-source-id: 3566b27