Add additional logging to help diagnose missing print
Summary:
For some reason sometimes it seems like the event bus
gets really backed-up in terms of sending events.
There's a lack of repro, and impossible to debug.
Hopefully this will give us some better logs/insights into what
event is causing this, and hence what listeners are possibly
causing the issue in the logs.

Reviewed By: sbalabanov

fbshipit-source-id: 10c4752