Add a ModuleMap HeaderMode for modular apple_library
Summary:
This adds a new `SYMLINK_TREE_WITH_MODULEMAP` on `HeaderMode`. This will be the headermode for when an `apple_library` is modular.

The modulename is derived from the links, if any links are present. Lets the new mode integrate with the currently available interfaces of `CxxPreprocessables`. It should be okay as a modulemap will only work if the headers are in the Apple style.
Closes https://github.com/facebook/buck/pull/1784

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: b58a9b0