Add an argument for scheme pre and post actions in Xcode
Summary:
Xcode currently allows you to execute a shell step when you start an stop executing in a scheme. I would like buck to be able to generate a project with this configured.

I have an implementation of this on an internal fork, it works but the interface isn't very satisfying. Although the rule configuration seems reasonable:
```
xcode_workspace_config(
    ...
    additional_scheme_actions = {
      'Build': {
          'Pre': ['echo start'],
          'Post': ['echo stop'],
      }
    }
    ...
)
```
The current argument type that is coerced seems too long
```
Optional<ImmutableMap<SchemeActionType, ImmutableMap<XCScheme.AdditionalActions, ImmutableList<String>>>>
```
Any suggestions on how to handle this? Is this acceptable in java? Should I try to achieve this outside of buck?
Closes https://github.com/facebook/buck/pull/1507

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: 980fb9c