Cache conflicting header check results on the PreprocessorDelegate
Summary:
Conflicting headers check was moved out of action graph creation
into the build phase. However, the change caused the check to be re-evaluated
whenever a file is compiled, rather than once per library. This change caches
the value on the PreprocessorDelegate, ensuring it only gets evaluated once per
library.

Additionally, add some documentation on what the conflicting header check is
and why it exists.

Additionally, change the interface slightly.

The actual effect of this change is small in terms of cpu time, but might
stress the GC somewhat since it has to create lots of somewhat large throwaway
HashMaps during build.

Reviewed By: bobyangyf

fbshipit-source-id: 3d22411