Remove method for test only from CachingBuildEngine
Summary:
This method was only used for tests.
This also indirectly relied oni by protected methods of ManifestRuleKeyManager,
and it's corresponding tests. And since VisibleForTesting relies on the test
being in the same package as the class, this couples the CachingBuildEngine tests
to being in the same package as ManifestRuleKeyManager.
Moving this method to be part of the test breaks the dependency

Reviewed By: ttsugriy

fbshipit-source-id: fc4532f