Move method used for test only from RemoteBuildRuleSynchronizer to RemoteBuildRulesynchronizerTestUtil
Summary:
This method was used by CachingBuildRuleTest. Since VisibleForTesting
relies on test and class to be in the same package, this forces
CachingBuildEngine and the RemoteBuildRulesynchronizer to be in the same
package, which isn't necessary.
We make a test util that exposes the test only method to all tests only.

Reviewed By: ttsugriy

fbshipit-source-id: a85b6b8