Fixed `buck publish` to stop incorrectly packaging direct dependencies sources.
The problem was the difference between how TraversedDeps is initialized for sources jar versus regular jar.

For the sources jar java_library Build rule is not available yet and set of all classpath dependencies was used as root and auto-included in sources jar.

This led to sources jar including superset of what it is supposed to contain.