Introduce BuildFileManifest.
Summary:
Passing a set of raw targets is very error prone (clients use indicies and perform casts) and inconvenient.
This change does not replace all usages since it's very large already, but
will be extended in follow up commits. The eventual goal is to use it everywhere
and use it to include globs metadata.

Reviewed By: styurin

fbshipit-source-id: 6af5d06