rust: add prefer_static_libs flag
Summary:
Add prefer_static_libs config flag, mostly so that shared builds can still statically link with the standard libraries.

Normally shared libraries are preferred when doing a shared build. When this config flag is set, then it will prefer static libraries. This is a tie-breaker - when doing a shared build, if both shared and static libraries are available, the static ones are preferred when this is set. If only one of shared or static is available, then its used regardless of this flag's setting.

Reviewed By: andrewjcg

fbshipit-source-id: c8883b2