Genrule files are added only if they're not already referenced
Summary:
Some files are showing as duplicates when they are referenced in apple_library and in genrules.
They also appeared in the group "Other" in addition to the group for the apple_library.
This patch will not add it to the group "Other" if it's already part of another target.

Reviewed By: ttsugriy

fbshipit-source-id: 796b39a