Enable IntelliJ source lookup for first party dependencies when used with `buck project`.
Summary:
This should allow 'goto definition' to work as you'd expect within a monorepo.

NOTE: Due to the way IntelliJ works, this only works if your sources have a proper "source root". This means that a class `a.b.c.D`, *must* live in `$ROOT/a/b/c/D.java`, for some value of `$ROOT`.

 ---

This diff works by updating the content that's generated in the XML files under `.idea/libraries/__*` with the proper source roots.

For instance, if a target in `foo/bar/...` depends on `foo/buck:baz_lib`, `buck project foo/bar/...` would produce the following library XML file for `foo/buck:baz_lib` (`foo/bar/.idea/libraries/__foo_buck_baz_lib.xml`):

```
<component name="libraryTable">
  <library name="//foo/buck:baz_lib">
    <CLASSES>
      <root url="jar://$PROJECT_DIR$/../../buck-out/dev/gen/foo/buck/lib__baz_lib__output/baz_lib.jar!/" />
    </CLASSES>
    <JAVADOC />
    <SOURCES />
  </library>
</component>
```

but now it looks like:

```
<component name="libraryTable">
  <library name="//foo/buck:baz_lib">
    <CLASSES>
      <root url="jar://$PROJECT_DIR$/../../buck-out/dev/gen/foo/buck/lib__baz_lib__output/baz_lib.jar!/" />
    </CLASSES>
    <JAVADOC />
    <SOURCES>
      <root url="file://$PROJECT_DIR$/../buck/src/main/java" />
    </SOURCES>
  </library>
</component>
```

The additional `SOURCES` root makes IntelliJ pick up the paths to the files.

NOTE: It doesn't actually work if you specify a directory that directly contains the source files as source roots. The source root must actually have a directory structure that actually matches the package of the class. `buck` will allow targets with sources that do not follow this convention, but if this convention is not followed, IntelliJ will just not work properly with the project.

Reviewed By: styurin

fbshipit-source-id: 6d3b087