Append "__" to gen/scratch root path for unflavored targets
Summary:
For unflavored targets, the gen/scratch root paths might be the same as
legitimate subdirectories of the base path. This has caused problems in
the past. A simple approach to resolving this that causes little
unexpected issues is to simply append "__" in such a case.

This change does that for DefaultOutputPathResolver. (it also caches the
computed Paths to better interact with BuckUnixFilesystem).

Reviewed By: ttsugriy

fbshipit-source-id: 6ddf2d2