Fix MultiThreadedBlobUploader
Summary:
Currently, a single upload request will trigger one job on the work
queue. That job might not actually be sufficient to drain the queue. This makes
the job requeue if there might be work to do.

Also, a job might run when the queues are empty (because another pulled
everything). This includes a fix to handle that correctly.

Reviewed By: ttsugriy

fbshipit-source-id: 2895a59