Fixed `buck publish` to stop incorrectly packaging direct dep srcs
Summary:
The problem was the difference between how TraversedDeps is initialized for sources jar versus regular jar.

For the sources jar java_library Build rule is not available yet and set of all classpath dependencies was used as root and auto-included in sources jar.

This led to sources jar including superset of what it is supposed to contain.
Closes https://github.com/facebook/buck/pull/1878

Reviewed By: styurin

Pulled By: styurin

fbshipit-source-id: 49f0277