Fixed file materialization issue where Future never completes on remote call error
Summary:
If during file materialization from a remote server an exception is thrown, the corresponding file materialization Future never completes, which means no progress is made until timeouts are hit.

New behavior is to set the future to an exception state when this happens, and also to log the error.

Reviewed By: selimekizoglu

fbshipit-source-id: 1a1c23d