Choose stacked build capacity tracking strategy based on configuration
Summary:
This adds a dist build config to enable greedy stacking. If the flag is set `GreedyMultiBuildCapacityTracker` is used instead of `DefaultMultiBuildCapacityTracker`.

The difference between default & greedy strategies is that default one asks build slave for available capacity and then commits for capacity based on the number of work units gathered from the coordinator which can result to overscheduling. Greedy one reserves all available capacity in advance instead of asking first and returns unused capacity if the number of work units got from the coordinator is less than reserved capacity (which could potentially starve other builds stacked on the build slave).

Reviewed By: aiked

fbshipit-source-id: 6c3ea35