Make .gitignore more tolerant of individual IntelliJ plugins.
Summary:
In general, IntelliJ plugins are permitted to add preferences to the
.idea/* directory.  Rather than blacklist the plugins that we don't
want to save, instead whitelist the plugins we *do* want, so that
developers who use custom plugins that save to .idea/ aren't pestered
to check in their prefs files.

As part of this, also add .idea/copyright/profiles_settings.xml and
point it to the (pre-existing) .idea/copyright/Facebook.xml to enable
copyright notices when creating new files.

Reviewed By: styurin

fbshipit-source-id: ff7b93e