Allow adding additional context to error messages based on config
Summary:
This adds the ability to postprocess parser errors in a user-configurable way. This takes a map of regexes that need to be transformed, and a replacement for each, which will be appended to the original error.

This can be helpful when assumptions are made in extension files that don't end up being true (e.g. assuming that a dependency exists), and the solution to the problem is well known (maybe make sure that the dependency is passed to a different function or attribute). This is a temporary solution until more introspection can be passed into our macro/build rule environment.

Reviewed By: ttsugriy

fbshipit-source-id: 21772fc