Remove unneeded stack frames from python_parser when running from a jar
Summary:
Depending on how we invoke the python parser, we don't always filter out parser frames from the stacktrace properly. The python dir that's passed in doesn't always correspond to where we're really running, since we sometimes run from a resource directory (and sometimes not, it appears that we switch depending on whether we're running from a jar or not).
Since we're not going to be changing the name of buck.py anytime soon, just filter out anything that matches a chunk of the the parser's relative path.

Reviewed By: ttsugriy

fbshipit-source-id: 841efe3