Add support for symlink inputs
Summary:
Many rules in buck have symlinks as their inputs. To properly support that, we
should materialize symlinks in the worker (and include the targets as part of
the inputs).

This updates FileTreeBuilder/IsolatedExecutionStrategy to handle symlinks
correctly (including directory symlinks).

This updates the protocols to include a SymlinkNode (this matche bazel's
symlink proposal here:
https://docs.google.com/document/d/1gnOYszitgrLVet3sQk-TKGqIcpkkDsc6aw-izoo-d64/preview).

Reviewed By: ttsugriy

fbshipit-source-id: d60e7a1