Add caching of filesystem derived information
Summary:
When serializing a lot of rules (especially c++) many different nodes
will have shared inputs. Computing the inputs trees requires knowing directory
contents, and whether files are directories. This gets expensive.

This change adds caching of that information in the serializer. This is safe
because the serializer is only used for the duration of a build, and we won't
be looking at inputs until after whatever has generated them has run.

Reviewed By: ttsugriy

fbshipit-source-id: 5b17beb