Emit friendlier message when compiler crashes in source-only mode
Summary: Sometimes what would be a nice error in normal mode results in a crash in source-only mode. This commit replaces the crash with an instruction to build the regular rule, in hopes that by fixing the nice errors it will stop the source-only mode crash. The full stack trace is still available in the log, so that we have the opportunity to investigate it and create a nicer error even in source-only mode.

Reviewed By: styurin

fbshipit-source-id: 163111b