Move BuildRule marker interfaces from rules.keys to rules
Summary:
These are marker interfaces to tag build rules with special properties.
This should belong with the rest of these interfaces and not in the rulekey
implementations in rules.keys.

These marker interfaces will be moved together later to a separate package,
but we need them to belong together for now to avoid circular dependencies during
moving

Reviewed By: ttsugriy

fbshipit-source-id: d5e5263