Improve error message for reporting errors
Summary:
There have been occasional errors in this code in CI and when that happens it's really strange to get just a single debug-level log line with a scary error message like "[Errno 2] No such file or directory". It also makes it hard to debug when you don't know where it's coming from.

In this diff, we add exception information to the log so that we can figure out what went wrong from looking at it. Since we now print a big scary stack trace, we also print out a more explicit error message explaining that the error is benign.

Reviewed By: ttsugriy

fbshipit-source-id: d587854