Rename Description and corresponding BuildRuleCreationContext to have WithTargetGraph as suffix
Summary:
With new Incremental action graph build engine, we will
no longer need TargetGraph in the BuildRuleCreationContext and
Description.
This allows renames the current Description and Context to mark it as
with TargetGraph so we can introduce a version that doesn't
have a dependency on TargetGraph and have them in a separate core.description package.

Reviewed By: ttsugriy

fbshipit-source-id: be07546