measure hit ratio based on requests, not rules
Summary:
Currently the UI is printing the cache ratio as a result of successful cache hits / number of rules. The problem with that is that un-cacheable rules are registering as cache hits. My goal is to change that and evaluate the ratio based on how many successful hits we had from the rules we could possibly get a hit.

Also a minor change on the UI to show the information on the same line as the rest of the info for the caching.

Reviewed By: cjhopman

fbshipit-source-id: 5166008