Revert "Revert "Enable IntelliJ source lookup for first party dependencies when used with `buck project`.""
Summary:
This reverts commit e0e7e3d3edb1862b48536b0d70ae834a69a454dd.
The changes in this commit are identical to that original one.

I'm stacking this commit so that it will be easier to review the changes on top of the reverted changes.

 ---

Turns out https://github.com/facebook/buck/issues/1886 was a scalability issue. Since, the logic is still ok, and useful for many people, this feature is going to be hidden behind a flag in a stacked commit.

Reviewed By: styurin

fbshipit-source-id: 8f7c1c2