Include native function name in the error message.
Summary:
Previously the error would be a very generic
```
The native module cannot be accessed from here. Wrap the function in a macro and call it from a BUCK file
```

now it's going to be a more user friendly:
```
Top-level invocations of read_config are not allowed in .bzl files. Wrap it in a macro and call it from a BUCK file.
```

Reviewed By: cjhopman

fbshipit-source-id: 41f31a4