Remove BuildRuleResolver field from StringWithMacrosConverter
Summary:
This is a stepping stone to fixing some lambda captures that are problematic in incremental action graph construction.

There are a few places where we store or capture `StringWithMacrosConverter`, which is problematic since the class has a `BuildRuleResolver` field. Removing the field, and adding a parameter to `convert` instead.

Reviewed By: cjhopman

fbshipit-source-id: a4fbf1a