Allow Build Errors to be augmented
Summary:
Errors from the build itself do not follow the same path as parse errors. That is its own issue, but for now, make sure we add error augmentations to build failures as well.

This also moves the augmentation class around and has to break up a rule in core/exceptions to prevent cycles.

Reviewed By: ttsugriy

fbshipit-source-id: 81dd0e9