Split BuildRuleResolver into a more limited interface and an ActionGraphBuilder
Summary:
BuildRuleResolver provides a lot of access to the underlying action
graph (including the ability to trigger creation of new rules and even
to directly add target->rule mappings).

Ideally, BuildRuleResolvers wouldn't really be needed in so many places,
but that's not the world we live in.

This change makes BuildRuleResolver a more limited interface
containing just the methods for looking up rules corresponding to
particular buildtargets.

The previous BuildRuleResolver is renamed to ActionGraphBuilder.

Reviewed By: bobyangyf

fbshipit-source-id: 3f51a6c