Add modulemap to headers for modular apple_library
Summary:
When an `apple_library` is marked as modular, add a modulemap to the generated headersymlinktree so other rules can import this library modularly.

Modules are always platform specific so we need to only have one platform specific headersymlinktree with a modulemap. We cannot forward the metadata to a CxxLibraryDescription as it will create two trees (platform specific and platform agnostic). This PR makes sure that if a library is modular, AppleLibraryDescription doesn't forward the `CxxLibraryDescription.MetadataType.CXX_PREPROCESSOR_INPUT` to a CxxLibraryDelegate but instead creates it itself.
Closes https://github.com/facebook/buck/pull/1874

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: 08ad8eb