Allow modular mixed apple_library
Summary:
This pr adds the functionality to support modular mixed objective-c/swift librariesusing the `-import-underlying-module` flag. The new swift metadata type `SWIFT_UNDERLYING_MODULE` refers to a build rule for a headersymlinktree that exposes the public Objective-C headers through a module with the same name, so that the swift compiler can read this using the import-underlying-module flag when compiling the swift sources for the library.

Based on having support for modular libraries in #1874.
Closes https://github.com/facebook/buck/pull/1875

Reviewed By: milend

Pulled By: milend

fbshipit-source-id: 371d4f3