Add --output-attribute
Summary: Add --output-attribute as an eventual replacement for --output-attributes. This is done because right now the nature of --output-attributes is to consume all bare arguments (that don't start with -). This includes the query itself, and any extra arguments if one does `%s //foo: //bar:`. This syntax lets one do `buck query --output-attribute '.*' %s //foo: //bar:` and makes it easier to go back in a terminal to modify the query without having to jump back over the ambiguous --output-attributes flag.

Reviewed By: ttsugriy

fbshipit-source-id: a3581c0