Allow integer 0-100 coverage ratio percentages
Summary:
As part of the migration to skylark, we're going to be removing the Float type from our rules. This is because skylark does not support floats. This updates the needed coverage rule to support integral percentages in addition to its existing behavior. Values over 1 will be percentages, values under 1 will be ratios.

Ratios provided to external test runners should remain unchanged and will still be floats, as that is part of our contract with them.

Reviewed By: ttsugriy

fbshipit-source-id: 4c0d93a