Fix possible race condition when saving binary to cache
Summary: When we uploading artifacts to buck, there is a possibility that the metadata will be uploaded before content. This might cause another worker to read metadata, fail to load content, and reupload the artifact (metadata + content). With high volume of reads a writes we can end up reloading the same artifact over and over again.

Reviewed By: ttsugriy

fbshipit-source-id: b3f06b7