TargetNodeToBuildRuleTransformer to take a ToolchainProvider instead of CellProvider
Summary:
The CellProvider is only used to lookup the Cell for the buildtarget and
then that's only used to lookup the ToolchainProvider. We shouldn't be
exposing all the information available from the CellProvider to these
transformers.

This also changes the actiongraph builders to only capture the ability
to lookup toolchain providers from the passed in cellprovider (these
should probably also be updated to not take a cellprovider).

Reviewed By: bobyangyf

fbshipit-source-id: 600ea01