Computation of the longest path in terms of time in the action graph of the test target
Summary:
Regarding the dependencies of each build rule, the longest path (critical path) is computed. I have two implementations: one is similar to the computation in the `DistBuildTrace` which is nonlinear, and other implementation that uses topological sort which is linear overall.
`duration` of each build rule is computed as difference between the finish time of `BuildRuleEvent.Finished` and the start time of `BuildRuleEvent.Started`.

Reviewed By: bobyangyf

fbshipit-source-id: b373ca1