Change MultiThreadedActionGraphBuilder.computeIfAbsent to complete the future.
Summary:
We may have cases where a future being completed calls on
`computeIfAbsent` to try to compute it self. We previously just
ran the lambda without setting the Future.
We should actually complete the future, verify that the result is consistent.

There is also minor perf wins here in avoiding recomputation in cases
where the same build target is required multiple times.

Reviewed By: ttsugriy

fbshipit-source-id: fbf2863