slightly reduce memory usage of HeaderPathNormalizer
Summary:
Compute absolute path of a SourcePath only once so we don't store two copies
into the HeaderPathNormalizer.

This might reduce memory usage of HeaderPathNormalizers by 25% or more. Remember
that HeaderPathNormalizers are weakly-held, and for cxx heavy repos with lots of headers and libraries, may take up significant heap
size between GC cycles (50%+). The smaller they are and the fewer times they are
GC'd the better.

Reviewed By: sbalabanov

fbshipit-source-id: b3dc996