Add support for concatenation in coercers
Summary:
Adds a new property to a coercer - ability to concatenate values. This will be used together with lists of selectable values.

Coercers for types that support concatenation will have `concat` method implemented while those that don't support should use the default implementation that returns `null` to indicate that the operation is not supported.

Typical example of where this functionality will be used:

```
rule(
  ...
  deps = [
    ...
  ] + select({
    ...
  }) + select({
    ...
  }),
)
```

In this example a list with selects contains three elements. When `select` statements are resolved to some values (lists in this case), the result list of computed by calling `concat` with three elements (one original list and two lists obtained from the resolution.)

Reviewed By: ttsugriy

fbshipit-source-id: 6591fd2