Introduce Selector and related classes
Summary:
This change introduces classes providing support for `select` functionality. These are the basic classes that allows us to represent `select` statements and preform resolution of conditions.

Typical example of `select`:

```
rule(
  ...
  deps = [
    "//a:non-selectable",
  ] + select({
    "//condition:a": ["//a:a"],
    "DEFAULT": ["//a:default"],
  }) + select({
    "//condition:b": ["//a:b"],
    "//condition:c": ["//a:c"],
    no_match_error = "Custom error with some details",
  }),
)
```

`Selector` class represents `select` statement which contains conditions, a message for a situation when no conditions match and flag to indicate that one of the conditions is default.

`SelectrList` represents a list of elements some of which maybe selectable. From the example above, a list has three elements, first element is a list with `//a:non-selectable` as a single element and two other elements as instances of `Selector`.

Conditions are represented as maps with keys (`SelectorKey`) either pointing to a build target or being a default key.

Non-default conditions point to targets that can provide `Selectable` (they implement `ProvidesSelectable`). `Selectable` is used by `SelectorListResolver` to evaluate conditions in `Selector`s and produce the result by concatenating resolved values. In the example above, if only `//condition:c` matches the configuration, the resolution would return a list consisting of three elements: `["//a:non-selectable", "//a:default", "//a:c"]`.

For more information about resolution see Bazel documentation: https://docs.bazel.build/versions/master/be/functions.html#select

Reviewed By: ttsugriy

fbshipit-source-id: dffca89