JsLibrary: Fix tests with streams
Summary:
`JsLibraryDescriptionTest` was using `peek(...).findAny()` in order to run assertions on a stream of values. The intention was to guarantee that assertions were run.

This method is broken, as it will short-circuit stream evaluation after running the first assertion, not asserting on the remaining values.

Here, we replace the system with a custom `Collector` that will consume the stream completely, and throw if the stream was empty, i.e. no assertions were run.

Reviewed By: cjhopman

fbshipit-source-id: cb2cccb