Add trace event for HeaderPathNormalizer
Summary:
This will help with debugging when these things are re-generated, and give
insight as to how much they contribute to overall build time.

In order to do this I am plumbing down the nearest BuildContext to
getHeaderPathNormalizer, instead of just the SourcePathResolver. This seems
semantically correct as generating HeaderPathNormalizers is something done
during the build.

Reviewed By: jtorkkola

fbshipit-source-id: 95539e1