Fix ImmutableTypeCoercer.traverse to properly handle immutable object and its fields
Summary:
Traversing an immutable object is broken because fields' coercers are used to traverse the immutable object instead of traversing fields only.

This change fixes the logic and also adds a step to traverse the whole immutable object with a given traverser.

This wasn't caught because swallowing exceptions which is fixed earlier.

Reviewed By: ttsugriy

fbshipit-source-id: 319340b